move bulk of logic out of main file
