+++ connector/connector_test.go
@@ -1,4 +1,4 @@
-package main
+package connector
 
 import "testing"
